chore: Improve getBillingCycle function in utils.php
Refactor the `getBillingCycle` function in `utils.php` to handle the case when the lookup date is the same as the current date. If they match, the function now modifies the lookup date to be one month ahead and returns the start and end dates of the billing cycle. This improves the accuracy of the billing cycle calculation.