Refactor TinyDebugger formatOutput method to support plain text output
The `formatOutput` method in the `TinyDebugger` trait has been refactored to support plain text output in addition to HTML. A new optional parameter `$asHTML` has been added to control the output format. When `$asHTML` is set to `false` or when running in CLI mode, the method will return a formatted plain text string instead of HTML. This improves the flexibility and usability of the debugging functionality.